+++ components/src/main/java/com/mercadolibre/android/andesui/radiobutton/align/AndesRadioButtonAlignInterface.kt
@@ -1,0 +1,7 @@
+package com.mercadolibre.android.andesui.radiobutton.align
+
+internal sealed class AndesRadioButtonAlignInterface
+
+internal object AndesRadioButtonLeft : AndesRadioButtonAlignInterface()
+
+internal object AndesRadioButtonRight : AndesRadioButtonAlignInterface()