+++ components/src/test/java/com/mercadolibre/android/andesui/radiobutton/AndesRadioButtonTest.kt
@@ -1,0 +1,297 @@
+package com.mercadolibre.android.andesui.radiobutton
+
+import android.os.Build
+import com.mercadolibre.android.andesui.BuildConfig
+import com.mercadolibre.android.andesui.R
+import com.mercadolibre.android.andesui.color.toAndesColor
+import com.mercadolibre.android.andesui.radiobutton.align.AndesRadioButtonAlign
+import com.mercadolibre.android.andesui.radiobutton.factory.AndesRadioButtonAttrs
+import com.mercadolibre.android.andesui.radiobutton.factory.AndesRadioButtonConfigurationFactory
+import com.mercadolibre.android.andesui.radiobutton.status.AndesRadioButtonStatus
+import com.mercadolibre.android.andesui.radiobutton.type.AndesRadioButtonType
+import org.junit.Assert.assertEquals
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.mockito.Mockito.spy
+import org.robolectric.RobolectricTestRunner
+import org.robolectric.RuntimeEnvironment
+import org.robolectric.annotation.Config
+
+@RunWith(RobolectricTestRunner::class)
+@Config(constants = BuildConfig::class, sdk = [Build.VERSION_CODES.LOLLIPOP])
+class AndesRadioButtonTest {
+
+    private var context = RuntimeEnvironment.application
+    private val configFactory = spy(AndesRadioButtonConfigurationFactory)
+    private lateinit var attrs: AndesRadioButtonAttrs
+
+    @Test
+    fun `RadioButton, Idle, Unselected, Border`() {
+        attrs = AndesRadioButtonAttrs(
+                andesRadioButtonAlign = AndesRadioButtonAlign.LEFT,
+                andesRadioButtonStatus = AndesRadioButtonStatus.UNSELECTED,
+                andesRadioButtonType = AndesRadioButtonType.IDLE,
+                andesRadioButtonText = "Andes checkbox"
+        )
+        val config = configFactory.create(attrs)
+        assertEquals(
+                config.type.type.borderColor(context, config.status),
+                R.color.andes_gray_250_solid.toAndesColor()
+        )
+    }
+
+    @Test
+    fun `RadioButton, Idle, Unselected, Background`() {
+        attrs = AndesRadioButtonAttrs(
+                andesRadioButtonAlign = AndesRadioButtonAlign.LEFT,
+                andesRadioButtonStatus = AndesRadioButtonStatus.UNSELECTED,
+                andesRadioButtonType = AndesRadioButtonType.IDLE,
+                andesRadioButtonText = "Andes checkbox"
+        )
+        val config = configFactory.create(attrs)
+        assertEquals(
+                config.type.type.backgroundColor(context, config.status),
+                R.color.andes_white.toAndesColor()
+        )
+    }
+
+    @Test
+    fun `RadioButton, Idle, Unselected, Icon color`() {
+        attrs = AndesRadioButtonAttrs(
+                andesRadioButtonAlign = AndesRadioButtonAlign.LEFT,
+                andesRadioButtonStatus = AndesRadioButtonStatus.UNSELECTED,
+                andesRadioButtonType = AndesRadioButtonType.IDLE,
+                andesRadioButtonText = "Andes checkbox"
+        )
+        val config = configFactory.create(attrs)
+        assertEquals(
+                config.type.type.iconColor(context, config.status),
+                R.color.andes_white.toAndesColor()
+        )
+    }
+
+    @Test
+    fun `RadioButton, Idle, Selected, Border`() {
+        attrs = AndesRadioButtonAttrs(
+                andesRadioButtonAlign = AndesRadioButtonAlign.LEFT,
+                andesRadioButtonStatus = AndesRadioButtonStatus.SELECTED,
+                andesRadioButtonType = AndesRadioButtonType.IDLE,
+                andesRadioButtonText = "Andes checkbox"
+        )
+        val config = configFactory.create(attrs)
+        assertEquals(
+                config.type.type.borderColor(context, config.status),
+                R.color.andes_blue_ml_500.toAndesColor()
+        )
+    }
+
+    @Test
+    fun `RadioButton, Idle, Selected, Background`() {
+        attrs = AndesRadioButtonAttrs(
+                andesRadioButtonAlign = AndesRadioButtonAlign.LEFT,
+                andesRadioButtonStatus = AndesRadioButtonStatus.SELECTED,
+                andesRadioButtonType = AndesRadioButtonType.IDLE,
+                andesRadioButtonText = "Andes checkbox"
+        )
+        val config = configFactory.create(attrs)
+        assertEquals(
+                config.type.type.backgroundColor(context, config.status),
+                R.color.andes_blue_ml_500.toAndesColor()
+        )
+    }
+
+    @Test
+    fun `RadioButton, Idle, Selected, Icon color`() {
+        attrs = AndesRadioButtonAttrs(
+                andesRadioButtonAlign = AndesRadioButtonAlign.LEFT,
+                andesRadioButtonStatus = AndesRadioButtonStatus.SELECTED,
+                andesRadioButtonType = AndesRadioButtonType.IDLE,
+                andesRadioButtonText = "Andes checkbox"
+        )
+        val config = configFactory.create(attrs)
+        assertEquals(
+                config.type.type.iconColor(context, config.status),
+                R.color.andes_white.toAndesColor()
+        )
+    }
+
+    @Test
+    fun `RadioButton, Disabled, Unselected, Border`() {
+        attrs = AndesRadioButtonAttrs(
+                andesRadioButtonAlign = AndesRadioButtonAlign.LEFT,
+                andesRadioButtonStatus = AndesRadioButtonStatus.UNSELECTED,
+                andesRadioButtonType = AndesRadioButtonType.DISABLED,
+                andesRadioButtonText = "Andes checkbox"
+        )
+        val config = configFactory.create(attrs)
+        assertEquals(
+                config.type.type.borderColor(context, config.status),
+                R.color.andes_gray_100_solid.toAndesColor()
+        )
+    }
+
+    @Test
+    fun `RadioButton, Disabled, Unselected, Background`() {
+        attrs = AndesRadioButtonAttrs(
+                andesRadioButtonAlign = AndesRadioButtonAlign.LEFT,
+                andesRadioButtonStatus = AndesRadioButtonStatus.UNSELECTED,
+                andesRadioButtonType = AndesRadioButtonType.DISABLED,
+                andesRadioButtonText = "Andes checkbox"
+        )
+        val config = configFactory.create(attrs)
+        assertEquals(
+                config.type.type.backgroundColor(context, config.status),
+                R.color.andes_white.toAndesColor()
+        )
+    }
+
+    @Test
+    fun `RadioButton, Disabled, Unselected, Icon color`() {
+        attrs = AndesRadioButtonAttrs(
+                andesRadioButtonAlign = AndesRadioButtonAlign.LEFT,
+                andesRadioButtonStatus = AndesRadioButtonStatus.UNSELECTED,
+                andesRadioButtonType = AndesRadioButtonType.DISABLED,
+                andesRadioButtonText = "Andes checkbox"
+        )
+        val config = configFactory.create(attrs)
+        assertEquals(
+                config.type.type.iconColor(context, config.status),
+                R.color.andes_gray_250_solid.toAndesColor()
+        )
+    }
+
+    @Test
+    fun `RadioButton, Disabled, Selected, Border`() {
+        attrs = AndesRadioButtonAttrs(
+                andesRadioButtonAlign = AndesRadioButtonAlign.LEFT,
+                andesRadioButtonStatus = AndesRadioButtonStatus.SELECTED,
+                andesRadioButtonType = AndesRadioButtonType.DISABLED,
+                andesRadioButtonText = "Andes checkbox"
+        )
+        val config = configFactory.create(attrs)
+        assertEquals(
+                config.type.type.borderColor(context, config.status),
+                R.color.andes_gray_100_solid.toAndesColor()
+        )
+    }
+
+    @Test
+    fun `RadioButton, Disabled, Selected, Background`() {
+        attrs = AndesRadioButtonAttrs(
+                andesRadioButtonAlign = AndesRadioButtonAlign.LEFT,
+                andesRadioButtonStatus = AndesRadioButtonStatus.SELECTED,
+                andesRadioButtonType = AndesRadioButtonType.DISABLED,
+                andesRadioButtonText = "Andes checkbox"
+        )
+        val config = configFactory.create(attrs)
+        assertEquals(
+                config.type.type.backgroundColor(context, config.status),
+                R.color.andes_gray_100_solid.toAndesColor()
+        )
+    }
+
+    @Test
+    fun `RadioButton, Disabled, Selected, Icon color`() {
+        attrs = AndesRadioButtonAttrs(
+                andesRadioButtonAlign = AndesRadioButtonAlign.LEFT,
+                andesRadioButtonStatus = AndesRadioButtonStatus.SELECTED,
+                andesRadioButtonType = AndesRadioButtonType.DISABLED,
+                andesRadioButtonText = "Andes checkbox"
+        )
+        val config = configFactory.create(attrs)
+        assertEquals(
+                config.type.type.iconColor(context, config.status),
+                R.color.andes_gray_250_solid.toAndesColor()
+        )
+    }
+
+    @Test
+    fun `RadioButton, Error, Unselected, Border`() {
+        attrs = AndesRadioButtonAttrs(
+                andesRadioButtonAlign = AndesRadioButtonAlign.LEFT,
+                andesRadioButtonStatus = AndesRadioButtonStatus.UNSELECTED,
+                andesRadioButtonType = AndesRadioButtonType.ERROR,
+                andesRadioButtonText = "Andes checkbox"
+        )
+        val config = configFactory.create(attrs)
+        assertEquals(
+                config.type.type.borderColor(context, config.status),
+                R.color.andes_red_500.toAndesColor()
+        )
+    }
+
+    @Test
+    fun `RadioButton, Error, Unselected, Background`() {
+        attrs = AndesRadioButtonAttrs(
+                andesRadioButtonAlign = AndesRadioButtonAlign.LEFT,
+                andesRadioButtonStatus = AndesRadioButtonStatus.UNSELECTED,
+                andesRadioButtonType = AndesRadioButtonType.ERROR,
+                andesRadioButtonText = "Andes checkbox"
+        )
+        val config = configFactory.create(attrs)
+        assertEquals(
+                config.type.type.backgroundColor(context, config.status),
+                R.color.andes_white.toAndesColor()
+        )
+    }
+
+    @Test
+    fun `RadioButton, Error, Unselected, Icon color`() {
+        attrs = AndesRadioButtonAttrs(
+                andesRadioButtonAlign = AndesRadioButtonAlign.LEFT,
+                andesRadioButtonStatus = AndesRadioButtonStatus.UNSELECTED,
+                andesRadioButtonType = AndesRadioButtonType.ERROR,
+                andesRadioButtonText = "Andes checkbox"
+        )
+        val config = configFactory.create(attrs)
+        assertEquals(
+                config.type.type.iconColor(context, config.status),
+                R.color.andes_white.toAndesColor()
+        )
+    }
+
+    @Test
+    fun `RadioButton, Error, Selected, Border`() {
+        attrs = AndesRadioButtonAttrs(
+                andesRadioButtonAlign = AndesRadioButtonAlign.LEFT,
+                andesRadioButtonStatus = AndesRadioButtonStatus.SELECTED,
+                andesRadioButtonType = AndesRadioButtonType.ERROR,
+                andesRadioButtonText = "Andes checkbox"
+        )
+        val config = configFactory.create(attrs)
+        assertEquals(
+                config.type.type.borderColor(context, config.status),
+                R.color.andes_red_500.toAndesColor()
+        )
+    }
+
+    @Test
+    fun `RadioButton, Error, Selected, Background`() {
+        attrs = AndesRadioButtonAttrs(
+                andesRadioButtonAlign = AndesRadioButtonAlign.LEFT,
+                andesRadioButtonStatus = AndesRadioButtonStatus.SELECTED,
+                andesRadioButtonType = AndesRadioButtonType.ERROR,
+                andesRadioButtonText = "Andes checkbox"
+        )
+        val config = configFactory.create(attrs)
+        assertEquals(
+                config.type.type.backgroundColor(context, config.status),
+                R.color.andes_white.toAndesColor()
+        )
+    }
+
+    @Test
+    fun `RadioButton, Error, Selected, Icon color`() {
+        attrs = AndesRadioButtonAttrs(
+                andesRadioButtonAlign = AndesRadioButtonAlign.LEFT,
+                andesRadioButtonStatus = AndesRadioButtonStatus.SELECTED,
+                andesRadioButtonType = AndesRadioButtonType.ERROR,
+                andesRadioButtonText = "Andes checkbox"
+        )
+        val config = configFactory.create(attrs)
+        assertEquals(
+                config.type.type.iconColor(context, config.status),
+                R.color.andes_white.toAndesColor()
+        )
+    }
+}